requirements.txt
# ── 코스닥 세력 매집 감지 시스템 의존성 ──────────────────────
#
# 설치 방법 (Windows, 컴파일러 없는 환경):
#   python -m pip install -r requirements.txt
#   python -m pip install pykrx --no-deps   ← pykrx는 별도 설치
#
# 이유: pykrx가 numpy<2.0을 요구하지만, 컴파일러 없이는 numpy 1.x
#       소스 빌드가 불가. --no-deps로 pykrx만 설치 후 아래 의존성 사용.

# ── pykrx 런타임 의존성 (--no-deps 보완) ─────────────────────
multipledispatch>=0.6.0
tqdm>=4.0.0
Deprecated>=1.2.13

# ── 데이터 처리 ────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.23.0        # numpy 2.x도 pykrx 런타임 호환 (기능상 문제 없음)

# ── 기술적 지표 (순수 Python, 컴파일 불필요) ──────────────────
ta>=0.11.0

# ── DART 공시 ─────────────────────────────────────────────────
OpenDartReader>=0.2.0

# ── 스케줄러 ──────────────────────────────────────────────────
APScheduler>=3.10.0

# ── 텔레그램 ──────────────────────────────────────────────────
requests>=2.28.0
# python-telegram-bot>=20.0   # 선택: 설치 시 주석 해제

# ── 리포트 ────────────────────────────────────────────────────
Jinja2>=3.1.0
matplotlib>=3.6.0

# ── 설정 ──────────────────────────────────────────────────────
PyYAML>=6.0

# ── 테스트 ────────────────────────────────────────────────────
pytest>=7.2.0
pytest-mock>=3.10.0
